Istore task 16 and 27(backend)

+++ server/iStore/src/main/java/com/iStore/iStore/service/OrderDetailServiceImpl.java
@@ -14,6 +14,7 @@ import org.springframework.stereotype.Service;
 
 import com.iStore.iStore.constants.ISTOREConstants;
 import com.iStore.iStore.constants.PaymentMode;
+import com.iStore.iStore.model.Contact;
 import com.iStore.iStore.model.GenericResponse;
 import com.iStore.iStore.model.Item;
 import com.iStore.iStore.model.OrderDetail;
@@ -21,6 +22,7 @@ import com.iStore.iStore.model.OrderTotal;
 import com.iStore.iStore.repository.ItemRepository;
 import com.iStore.iStore.repository.OrderDetailRepository;
 import com.iStore.iStore.util.DateHelper;
+import com.iStore.iStore.util.RegexValidator;
 import com.iStore.iStore.util.UniqueDatesHelper;
 
 @Service
@@ -35,6 +37,15 @@ public class OrderDetailServiceImpl implements OrderDetailService {
 	@Override
 	public OrderDetail create(OrderDetail entity) {
 		try {
+			Contact contact = entity.getContact();
+			String name = contact.getName().trim();
+			String phone = contact.getPhone().trim();
+			if (name != null && !RegexValidator.isValidName(name)) {
+				throw new ValidationException("please enter valid name");
+			}
+			if (phone != null && !RegexValidator.isValidPhoneNumber(phone)) {
+				throw new ValidationException("please enter valid phone number");
+			}
 			productService.deleteInventory(entity.getItems());
 			List<Item> list = entity.getItems();
 			for (Item i : list) {
